services/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.20

# Database
sqlalchemy>=2.0.0

# PostgreSQL driver (optional — only needed if using PostgreSQL instead of SQLite)
# psycopg[binary]>=3.1.0

# ML / Data
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.5.0
joblib>=1.4.0

# AI / STT (optional — services fall back gracefully when keys are absent)
anthropic>=0.49.0
openai>=1.78.0

# LangChain (PDSAIBot hybrid agent — falls back gracefully without API key)
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0

# NLU (PDSAIBot embedded engine — PyYAML reads rasa/nlu.yml)
PyYAML>=6.0.0

# Twilio (optional — IVR only)
twilio>=9.4.6

# TTS — multilingual speech synthesis (requires internet; falls back gracefully)
gtts>=2.5.0

# Visualization (anomaly service)
matplotlib>=3.9.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Utilities
websockets>=14.2
httpx>=0.28.1
